made changes from feedback - colour changes, moved component location and name
